Выносит логику отрисовки в контроллер
